Clean up notification components
Quite a bit of duplicated code for the base card.
