+++ src/lib/components/notifications/MediaNotification.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
   import { Link } from "svelte-navigator";
-  import { format } from "timeago.js";
   import { NotificationType, type MediaFragment, type AiringNotification, type RelatedMediaAdditionNotification, type MediaDataChangeNotification } from "@anilist/graphql";
   import { hexToRgb } from "$lib/util";
+  import NotificationContainer from "./NotificationContainer.svelte";
 
   export let notification: Pick<AiringNotification | RelatedMediaAdditionNotification | MediaDataChangeNotification, "createdAt" | "type"> & {
     context?: string,
@@ -11,20 +11,13 @@
     media: Pick<MediaFragment, "id" | "title" | "img">,
   };
   export let unread = false;
-
-  const creationDate = new Date(notification.createdAt * 1000);
-  let displayDate = format(creationDate);
-  displayDate = displayDate.substring(0, displayDate.indexOf(" ") + 2).replace(" ", ""); 
 </script>
 
-<div 
-  class="relative p-2 flex items-center bg-foreground rounded-md border-r-4 { unread ? "border-r-accent" : "border-r-transparent" }"
-  style="--color-variable:{hexToRgb(notification.media.img.color) || "--color-accent"}"
->
+<NotificationContainer {unread} createdAt={notification.createdAt}>
   <Link to="/media/{notification.media.id}" class="flex-none">
     <img src={notification.media.img.large} alt="Key visual" class="w-12 mr-4 object-center object-cover aspect-[3/4]" />
   </Link>
-  <Link to="/media/{notification.media.id}" class="line-clamp-3 mr-6 flex-1">
+  <Link to="/media/{notification.media.id}" class="line-clamp-3 mr-6 flex-1" --color-variable={hexToRgb(notification.media.img.color) || "--color-accent"}>
     {#if notification.type === NotificationType.AIRING}
       {notification.contexts[0]}
       {notification.episode}
@@ -36,7 +29,4 @@
       {notification.context}
     {/if}
   </Link>
-  <time class="absolute top-1 right-2 text-xs" datetime={creationDate.toISOString()} title={creationDate.toLocaleString()}>
-    {displayDate}
-  </time>
-</div>
+</NotificationContainer>
