+++ src/lib/components/notifications/ThreadNotification.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
-  import { format } from "timeago.js";
   import type { ThreadFragment, UserFragment, ThreadLikeNotification, ThreadCommentLikeNotification, ThreadCommentReplyNotification, ThreadCommentMentionNotification, ThreadCommentSubscribedNotification } from "@anilist/graphql";
+  import NotificationContainer from "./NotificationContainer.svelte";
 
   export let notification: Pick<ThreadLikeNotification | ThreadCommentLikeNotification | ThreadCommentReplyNotification | ThreadCommentMentionNotification | ThreadCommentSubscribedNotification, "createdAt" | "context"> & {
     commentId?: number,
@@ -8,10 +8,6 @@
     user?: UserFragment
   };
   export let unread = false;
-
-  const creationDate = new Date(notification.createdAt * 1000);
-  let displayDate = format(creationDate);
-  displayDate = displayDate.substring(0, displayDate.indexOf(" ") + 2).replace(" ", ""); 
 
   function getContentUrl(): string {
     if (notification.thread.url === "#")
@@ -24,7 +20,7 @@
   }
 </script>
 
-<div class="relative p-2 flex items-center bg-foreground rounded-md border-r-4 { unread ? "border-r-accent" : "border-r-transparent" }">
+<NotificationContainer {unread} createdAt={notification.createdAt}>
   <a href={notification.user.url} target="_blank" class="flex-none">
     <img src={notification.user.img.large} alt="User profile" class="w-12 mr-4 object-center object-cover aspect-square rounded-full" />
     <!-- Merged -->
@@ -36,7 +32,4 @@
       {notification.thread.title}
     </span>
   </a>
-  <time class="absolute top-1 right-2 text-xs" datetime={creationDate.toISOString()} title={creationDate.toLocaleString()}>
-    {displayDate}
-  </time>
-</div>
+</NotificationContainer>
